refactor: Moved repeated functions into StepRunner
