+++ src/popper/runner.py
@@ -21,7 +21,7 @@ class WorkflowRunner(object):
         self._is_resman_module_loaded = False
 
     def _load_resman_module(self):
-        """dynamically load resource manager module"""
+        """dynamically load resource manager module."""
         resman_mod_name = f"popper.runner_{self._config.resman_name}"
         resman_spec = importlib.util.find_spec(resman_mod_name)
         if not resman_spec:
@@ -166,7 +166,7 @@ class WorkflowRunner(object):
         log.info("Workflow finished successfully.")
 
     def _step_runner(self, engine_name, step):
-        """Factory of singleton runners"""
+        """Factory of singleton runners."""
         if not self._is_resman_module_loaded:
             self._load_resman_module()
 
@@ -229,6 +229,93 @@ class StepRunner(object):
                 }
             )
         return step_env
+
+    def _get_build_info(self, step):
+
+        """Parses the `uses` attribute and returns build information needed.
+            Args:
+                step(dict): dict with step data
+            Returns:
+                (str, str, str, str): bool (build), image, tag, Dockerfile
+            """
+        build = True
+        img = None
+        build_ctx_path = None
+        img_full = None
+        tag = None
+
+        if "docker://" in step.uses:
+            img_full = step.uses
+            img = step.uses.replace("docker://", "")
+            if ":" in img:
+                (img, tag) = img.split(":")
+            else:
+                tag = "latest"
+            build = False
+
+        elif "./" in step.uses:
+            img_full = f'{pu.sanitized_name(step.id, "step")}'
+            img = img_full.lower()
+            tag = self._config.git_sha_short if self._config.git_sha_short else "na"
+            build_ctx_path = os.path.join(self._config.workspace_dir, step.uses)
+
+        else:
+            _, service, user, repo, step_dir, version = scm.parse(step.uses)
+            wf_cache_dir = os.path.join(self._config.cache_dir, self._config.wid)
+            repo_dir = os.path.join(wf_cache_dir, service, user, repo)
+            img_full = f"{user}/{repo}".lower()
+            img = img_full
+            tag = version
+            build_ctx_path = os.path.join(repo_dir, step_dir)
+
+        return (build, img_full, img, tag, build_ctx_path)
+
+    def _update_with_engine_config(self, container_args):
+
+        """Given container arguments, it extends it so it includes options
+        obtained from the popper.config.Config.engine_opts property.
+        """
+        update_with = self._config.engine_opts
+        if not update_with:
+            return
+
+        if container_args.get("volumes"):
+            container_args["volumes"] = [
+                *container_args["volumes"],
+                *update_with.get("volumes", list()),
+            ]
+        if container_args.get("bind"):
+            container_args["bind"] = [
+                *container_args["bind"],
+                *update_with.get("bind", list()),
+            ]
+
+        for k, v in update_with.get("environment", dict()).items():
+            container_args["environment"].update({k: v})
+
+        for k, v in update_with.items():
+            if k not in container_args.keys():
+                container_args[k] = update_with[k]
+
+    def _get_container_kwargs(self, step, img, name):
+        args = {
+            "image": img,
+            "command": list(step.args),
+            "name": name,
+            "volumes": [f"{self._config.workspace_dir}:/workspace:Z",],
+            "working_dir": step.dir if step.dir else "/workspace",
+            "environment": self._prepare_environment(step),
+            "entrypoint": step.runs if step.runs else None,
+            "detach": not self._config.pty,
+            "tty": self._config.pty,
+            "stdin_open": self._config.pty,
+        }
+
+        self._update_with_engine_config(args)
+        args.update(step.options)
+        log.debug(f"container args: {pu.prettystr(args)}\n")
+
+        return args
 
     def stop_running_tasks(self):
         raise NotImplementedError("Needs implementation in derived classes.")
