+++ src/popper/runner_host.py
@@ -172,42 +172,8 @@ class DockerRunner(StepRunner):
             log.info(f"Stopping container {c.name}")
             c.stop()
 
-    def _get_build_info(self, step):
-        """Parses the `uses` attribute and returns build information needed.
-
-        Args:
-            step(dict): dict with step data
-
-        Returns:
-            (str, str, str, str): bool (build), image, tag, Dockerfile
-        """
-        build = True
-        img = None
-        build_ctx_path = None
-
-        if "docker://" in step.uses:
-            img = step.uses.replace("docker://", "")
-            if ":" in img:
-                (img, tag) = img.split(":")
-            else:
-                tag = "latest"
-            build = False
-        elif "./" in step.uses:
-            img = pu.sanitized_name(step.id, "step").lower()
-            tag = self._config.git_sha_short if self._config.git_sha_short else "na"
-            build_ctx_path = os.path.join(self._config.workspace_dir, step.uses)
-        else:
-            _, service, user, repo, step_dir, version = scm.parse(step.uses)
-            wf_cache_dir = os.path.join(self._config.cache_dir, self._config.wid)
-            repo_dir = os.path.join(wf_cache_dir, service, user, repo)
-            img = f"{user}/{repo}".lower()
-            tag = version
-            build_ctx_path = os.path.join(repo_dir, step_dir)
-
-        return (build, img, tag, build_ctx_path)
-
     def _create_container(self, cid, step):
-        build, img, tag, build_ctx_path = self._get_build_info(step)
+        build, _, img, tag, build_ctx_path = self._get_build_info(step)
 
         if build:
             log.info(f"[{step.id}] docker build {img}:{tag} {build_ctx_path}")
@@ -233,6 +199,10 @@ class DockerRunner(StepRunner):
 
         container_args = self._get_container_kwargs(step, f"{img}:{tag}", cid)
 
+        if "volumes" not in container_args:
+            container_args["volumes"] = []
+        container_args["volumes"].insert(1, "/var/run/docker.sock:/var/run/docker.sock")
+
         log.debug(f"Container args: {container_args}")
 
         msg = f"[{step.id}] docker create name={cid}"
@@ -247,29 +217,6 @@ class DockerRunner(StepRunner):
 
         return container
 
-    def _get_container_kwargs(self, step, img, name):
-        args = {
-            "image": img,
-            "command": list(step.args),
-            "name": name,
-            "volumes": [
-                f"{self._config.workspace_dir}:/workspace",
-                "/var/run/docker.sock:/var/run/docker.sock",
-            ],
-            "working_dir": step.dir if step.dir else "/workspace",
-            "environment": self._prepare_environment(step),
-            "entrypoint": step.runs if step.runs else None,
-            "detach": not self._config.pty,
-            "tty": self._config.pty,
-            "stdin_open": self._config.pty,
-        }
-
-        self._update_with_engine_config(args)
-        args.update(step.options)
-        log.debug(f"container args: {pu.prettystr(args)}\n")
-
-        return args
-
     def _find_container(self, cid):
         """Check whether the container exists."""
         containers = self._d.containers.list(all=True, filters={"name": cid})
@@ -280,25 +227,6 @@ class DockerRunner(StepRunner):
             return filtered_containers[0]
 
         return None
-
-    def _update_with_engine_config(self, container_args):
-        """Given container arguments, it extends it so it includes options
-        obtained from the popper.config.Config.engine_opts property.
-        """
-        update_with = self._config.engine_opts
-        if not update_with:
-            return
-
-        container_args["volumes"] = [
-            *container_args["volumes"],
-            *update_with.get("volumes", list()),
-        ]
-        for k, v in update_with.get("environment", dict()).items():
-            container_args["environment"].update({k: v})
-
-        for k, v in update_with.items():
-            if k not in container_args.keys():
-                container_args[k] = update_with[k]
 
 
 class SingularityRunner(StepRunner):
@@ -390,50 +318,11 @@ class SingularityRunner(StepRunner):
         os.chdir(pwd)
         SingularityRunner.lock.release()
 
-    def _get_build_info(self, step):
-        build = True
-        img = None
-        build_ctx_path = None
-
-        if (
-            "docker://" in step.uses
-            or "shub://" in step.uses
-            or "library://" in step.uses
-        ):
-            img = step.uses
-            build = False
-
-        elif "./" in step.uses:
-            img = f'{pu.sanitized_name(step.id, "step")}'
-            build_ctx_path = os.path.join(self._config.workspace_dir, step.uses)
-        else:
-            _, service, user, repo, step_dir, version = scm.parse(step.uses)
-            wf_cache_dir = os.path.join(self._config.cache_dir, self._config.wid)
-            repo_dir = os.path.join(wf_cache_dir, service, user, repo)
-            img = f"{user}/{repo}".lower()
-            build_ctx_path = os.path.join(repo_dir, step_dir)
-
-        return (build, img, build_ctx_path)
-
     def _setup_singularity_cache(self):
         self._singularity_cache = os.path.join(
             self._config.cache_dir, "singularity", self._config.wid
         )
         os.makedirs(self._singularity_cache, exist_ok=True)
-
-    def _update_with_engine_config(self, container_args):
-        update_with = self._config.engine_opts
-        if not update_with:
-            return
-
-        container_args["bind"] = [
-            *container_args["bind"],
-            *update_with.get("bind", list()),
-        ]
-
-        for k, v in update_with.items():
-            if k not in container_args.keys():
-                container_args[k] = update_with[k]
 
     def _get_container_options(self):
         container_args = {
@@ -458,7 +347,12 @@ class SingularityRunner(StepRunner):
         return options
 
     def _create_container(self, step, cid):
-        build, image, build_ctx_path = self._get_build_info(step)
+        build, image, _, _, build_ctx_path = self._get_build_info(step)
+
+        if "shub://" in step.uses or "library://" in step.uses:
+            build = False
+            image = step.uses
+            build_ctx_path = None
 
         if build:
             log.info(f"[{step.id}] singularity build {cid} {build_ctx_path}")
