+++ src/test/test_runner.py
@@ -4,6 +4,7 @@ import shutil
 
 from unittest.mock import patch
 
+import popper.scm as scm
 from popper.cli import log
 from popper.config import ConfigLoader
 from popper.parser import WorkflowParser
@@ -139,3 +140,151 @@ class TestStepRunner(PopperTest):
             }
             self.assertDictEqual(expected, env)
             os.environ.pop("A")
+
+    def test_get_build_info(self):
+        step = Box(
+            {"uses": "popperized/bin/sh@master", "args": ["ls"], "id": "one",},
+            default_box=True,
+        )
+        with StepRunner() as r:
+            build, _, img, tag, build_ctx_path = r._get_build_info(step)
+            self.assertEqual(build, True)
+            self.assertEqual(img, "popperized/bin")
+            self.assertEqual(tag, "master")
+            self.assertTrue(f"{os.environ['HOME']}/.cache/popper" in build_ctx_path)
+            self.assertTrue("github.com/popperized/bin/sh" in build_ctx_path)
+
+            step = Box(
+                {
+                    "uses": "docker://alpine:3.9",
+                    "runs": ["sh", "-c", "echo $FOO > hello.txt ; pwd"],
+                    "env": {"FOO": "bar"},
+                    "id": "1",
+                },
+                default_box=True,
+            )
+        with StepRunner() as r:
+            build, _, img, tag, build_sources = r._get_build_info(step)
+            self.assertEqual(build, False)
+            self.assertEqual(img, "alpine")
+            self.assertEqual(tag, "3.9")
+            self.assertEqual(build_sources, None)
+
+        step = Box({"uses": "./", "args": ["ls"], "id": "one",}, default_box=True,)
+        conf = ConfigLoader.load(workspace_dir="/tmp")
+        with StepRunner(config=conf) as r:
+            build, _, img, tag, build_ctx_path = r._get_build_info(step)
+            self.assertEqual(build, True)
+            self.assertEqual(img, "popper_one_step")
+            self.assertEqual(tag, "na")
+            self.assertEqual(build_ctx_path, f"{os.path.realpath('/tmp')}/./")
+
+        # test within a git repo
+        repo = self.mk_repo()
+        conf = ConfigLoader.load(workspace_dir=repo.working_dir)
+        with StepRunner(config=conf) as r:
+            build, _, img, tag, build_ctx_path = r._get_build_info(step)
+            self.assertEqual(build, True)
+            self.assertEqual(img, "popper_one_step")
+            self.assertEqual(tag, scm.get_sha(repo, short=7))
+            self.assertEqual(build_ctx_path, f"{os.path.realpath(repo.working_dir)}/./")
+
+        step = Box(
+            {
+                "uses": "docker://alpine:3.9",
+                "runs": ["sh", "-c", "echo $FOO > hello.txt ; pwd"],
+                "env": {"FOO": "bar"},
+                "name": "1",
+            },
+            default_box=True,
+        )
+        with StepRunner() as r:
+            build, img_full, _, _, build_ctx_path = r._get_build_info(step)
+            self.assertEqual(build, False)
+            self.assertEqual(img_full, "docker://alpine:3.9")
+            self.assertEqual(build_ctx_path, None)
+
+    def test_get_container_kwargs(self):
+        step = Box(
+            {
+                "uses": "popperized/bin/sh@master",
+                "args": ["ls"],
+                "id": "one",
+                "dir": "/tmp/",
+                "options": {"ports": {"8888/tcp": 8888}},
+            },
+            default_box=True,
+        )
+
+        config_dict = {
+            "engine": {
+                "name": "docker",
+                "options": {
+                    "privileged": True,
+                    "hostname": "popper.local",
+                    "domainname": "www.example.org",
+                    "volumes": ["/path/in/host:/path/in/container"],
+                    "environment": {"FOO": "bar"},
+                },
+            },
+        }
+
+        config = ConfigLoader.load(
+            config_file=config_dict, workspace_dir="/path/to/workdir"
+        )
+
+        with StepRunner(config=config) as r:
+            args = r._get_container_kwargs(step, "alpine:3.9", "container_a")
+            self.assertEqual(
+                args,
+                {
+                    "image": "alpine:3.9",
+                    "command": ["ls"],
+                    "name": "container_a",
+                    "volumes": [
+                        "/path/to/workdir:/workspace:Z",
+                        "/path/in/host:/path/in/container",
+                    ],
+                    "working_dir": "/tmp/",
+                    "environment": {"FOO": "bar"},
+                    "entrypoint": None,
+                    "detach": True,
+                    "stdin_open": False,
+                    "tty": False,
+                    "privileged": True,
+                    "hostname": "popper.local",
+                    "domainname": "www.example.org",
+                    "ports": {"8888/tcp": 8888},
+                },
+            )
+
+        # check container kwargs when pty is enabled
+        config = ConfigLoader.load(
+            config_file=config_dict, workspace_dir="/path/to/workdir", pty=True
+        )
+
+        with StepRunner(config=config) as r:
+            args = r._get_container_kwargs(step, "alpine:3.9", "container_a")
+
+            self.assertEqual(
+                args,
+                {
+                    "image": "alpine:3.9",
+                    "command": ["ls"],
+                    "name": "container_a",
+                    "volumes": [
+                        "/path/to/workdir:/workspace:Z",
+                        "/path/in/host:/path/in/container",
+                    ],
+                    "working_dir": "/tmp/",
+                    "environment": {"FOO": "bar"},
+                    "entrypoint": None,
+                    "detach": False,
+                    "stdin_open": True,
+                    "tty": True,
+                    "privileged": True,
+                    "hostname": "popper.local",
+                    "domainname": "www.example.org",
+                    "ports": {"8888/tcp": 8888},
+                },
+            )
