+++ src/test/test_runner_slurm.py
@@ -218,7 +218,7 @@ class TestSlurmDockerRunner(unittest.TestCase):
             expected = (
                 "docker create"
                 " --name container_name"
-                " --workdir /workspace"
+                " --workdir /workspace:Z"
                 " -v /w:/workspace"
                 " -v /var/run/docker.sock:/var/run/docker.sock"
                 "   foo:1.9 -two -flags"
@@ -247,7 +247,7 @@ class TestSlurmDockerRunner(unittest.TestCase):
 
             expected = (
                 "docker create --name container_name "
-                "--workdir /workspace "
+                "--workdir /workspace:Z "
                 "-v /w:/workspace "
                 "-v /var/run/docker.sock:/var/run/docker.sock "
                 "-v /path/in/host:/path/in/container "
@@ -304,7 +304,7 @@ class TestSlurmDockerRunner(unittest.TestCase):
             expected = f"""#!/bin/bash
 docker rm -f popper_1_{config.wid} || true
 docker build -t popperized/bin:master {os.environ['HOME']}/.cache/popper/{config.wid}/github.com/popperized/bin/sh
-docker create --name popper_1_{config.wid} --workdir /workspace --entrypoint cat -v /w:/workspace -v /var/run/docker.sock:/var/run/docker.sock -v /path/in/host:/path/in/container -e FOO=bar   --privileged --hostname popper.local --domainname www.example.org popperized/bin:master README.md
+docker create --name popper_1_{config.wid} --workdir /workspace --entrypoint cat -v /w:/workspace:Z -v /var/run/docker.sock:/var/run/docker.sock -v /path/in/host:/path/in/container -e FOO=bar   --privileged --hostname popper.local --domainname www.example.org popperized/bin:master README.md
 docker start --attach popper_1_{config.wid}"""
             # fmt: on
             actual = f.read()
